move implementation of getting cost to controller
